fix compiler warnings

+++ findp/MikeHash.h
@@ -35,7 +35,7 @@ HT*		MikeHT_Init		(DWORD Entries);
 DWORD	MikeHT_Free		(HT *ht);
 
 //BOOL	MikeHT_Insert	(HT *ht, LPWSTR Key, LONGLONG Val);
-BOOL    MikeHT_Insert2  (HT* ht, LPCWSTR Key, const size_t KeyLen, LONGLONG Val);
+BOOL    MikeHT_Insert2  (HT* ht, LPCWSTR Key, const DWORD KeyLen, LONGLONG Val);
 BOOL    MikeHT_Get		(HT *ht, LPCWSTR Key, LONGLONG *Val);
 DWORD   MikeHT_ForEach  (HT *ht, KeyValCallback KeyValCallback, HT_STATS *stats, LPVOID context);
